+++ R/data.R
@@ -23,9 +23,9 @@
 "patient"
 
 
-#' Time-depedent covariates of the \code{\link{patient}} data set.
+#' Time-dependent covariates of the \code{\link{patient}} data set.
 #'
-#' This data set containt the time-dependent covariates (TDCs) for the \code{\link{patient}}
+#' This data set contains the time-dependent covariates (TDCs) for the \code{\link{patient}}
 #' data set. Note that nutrition was protocoled for at most 12 days after
 #' ICU admission. The data set includes:
 #' \describe{
@@ -45,14 +45,14 @@
 #' This is data simulated using the \code{\link[pammtools]{sim_pexp}} function.
 #' It contains two time-constant and two time-dependent covariates (observed
 #' on different exposure time grids). The code used for simulation is
-#' contained in the exmaples of \code{?sim_pexp}.
+#' contained in the examples of \code{?sim_pexp}.
 #'
 "simdf_elra"
 
 
 #' Stomach area tumor data
 #'
-#' Information on patients treated for a cancer desease
+#' Information on patients treated for a cancer disease
 #' located in the stomach area.
 #' The data set includes:
 #' \describe{
@@ -68,3 +68,31 @@
 #' }
 #'
 "tumor"
+
+
+
+#' Time until extubation
+#'
+#' This is a preprocessed
+#' subset of the data discussed in Heyard, et. al 2018 (and provided in a slightly
+#' different format as \code{VAP_data} in the package \code{TBFmultinomial}).
+#' In this package, the data is split in two parts, \code{extub_event} contains
+#' time-to-event data and time-constant covariates and \code{extub_tdc} contains
+#' the information on the time-dependent covariate SOFA score.
+#' The data contains the following variables:
+#' \describe{
+#'  \item{ID}{Unique patient ID}
+#'  \item{gender}{The patients' gender}
+#'  \item{type}{Type of admission, either \code{Medical} or \code{Surgical}}
+#'  \item{SAPSadmission}{SAPS score at admission}
+#'  \item{time}{Time (days) until extubation}
+#'  \item{extubation}{0 = no extubation/censoring, 1 = extubation}
+#'  \item{day}{Exposure time, i.e., time at which the SOFA score was observed}
+#' \item{SOFA}{The SOFA score at respective \code{day}s}
+#'}
+#' @aliases extub_tdc
+"extub_event"
+
+#' @inherit extub_event
+#' @rdname extub_event
+"extub_tdc"