+++ R/interval-information.R
@@ -116,7 +116,6 @@ get_intervals <- function(x, times, ...) {
 
 #' @inherit get_intervals
 #' @inheritParams base::findInterval
-#' @seealso \code{\link[base]{findInterval}}
 #' @rdname get_intervals
 #' @export
 get_intervals.default <- function(
@@ -172,7 +171,7 @@ ped_info <- function(ped) {
     bind_cols(
       int_df %>% slice(rep(seq_len(nrow(int_df)), times = nrow(sdf))),
       sdf %>% slice(rep(seq_len(nrow(sdf)), each = nrow(int_df)))) %>%
-      grouped_df(vars = groups(sdf))
+      grouped_df(vars = group_vars(sdf))
   }
 
 }
@@ -180,14 +179,15 @@ ped_info <- function(ped) {
 #' Extract risk set information for each interval.
 #'
 #' The columns \code{ped_riskset, ped_events, ped_censored} provide the
-#' size of the riskset at the beginning of each interval as well as the number
-#' of events and censorings that occured in the interval, respectively.
+#' size of the risk set at the beginning of each interval as well as the number
+#' of events and censorings that occurred in the interval, respectively.
 #'
 #' @inheritParams ped_info
 #' @import checkmate dplyr
 #' @examples
 #' ped <- tumor[1:4,] %>% as_ped(Surv(days, status)~ .)
 #' riskset_info(ped)
+#' @keywords internal
 #' @export
 #' @return A data frame with one row for each interval in \code{ped}.
 #' @seealso \code{\link[pammtools]{int_info}}, \code{\link[pammtools]{sample_info}}
@@ -198,10 +198,10 @@ riskset_info <- function(ped) {
   censored <- ped %>% group_by(id, add = TRUE) %>%
     arrange(.data$tend) %>% slice(n()) %>%
     filter(.data$ped_status == 0) %>% ungroup(id) %>%
-    grouped_df(vars = c(groups(ped), as.symbol("interval"))) %>%
+    grouped_df(vars = c(group_vars(ped), "interval")) %>%
     summarize(ped_censored = n())
 
-  join_vars <- unlist(c(sapply(groups(ped), deparse), "interval"))
+  join_vars <- c(group_vars(ped), "interval")
   ped %>% group_by(.data$interval, add = TRUE) %>%
     summarize(
       ped_riskset = n(),