+++ R/viz-elra.R
@@ -7,21 +7,17 @@
 #' @importFrom rlang quos
 #'
 #' @inheritParams make_newdata
+#' @param data Data used to fit the \code{model}.
 #' @param model A suitable model object which will be used to estimate the
 #' partial effect of \code{term}.
+#' @param term A character string indicating the model term for which partial
+#' effects should be plotted.
 #' @param reference If specified, should be a list with covariate value pairs,
-#' e.g. \code{list(x1 = 1, x=50)}. The calculated partial effect will be relative
+#' e.g. \code{list(x1 = 1, x2=50)}. The calculated partial effect will be relative
 #' to an observation specified in \code{reference}.
+#' @param ci Logical. Indicates if confidence intervals for the \code{term}
+#' of interest should be calculated/plotted. Defaults to \code{TRUE}.
 #' @export
-#' @examples
-#' ped <- tumor[1:200, ] %>% as_ped(Surv(days, status)~.)
-#' model <- mgcv::gam(ped_status~s(tend) + s(age, by = complications), data=ped,
-#'   family = poisson(), offset=offset)
-#' make_newdata(ped, age = seq_range(age, 20), complications = levels(complications))
-#' gg_slice(ped, model, "age", age=seq_range(age, 20), complications=levels(complications))
-#' gg_slice(ped, model, "age", age=seq_range(age, 20), complications=levels(complications),
-#'   reference=list(age = 50))
-#' @keywords internal
 gg_partial <- function(data, model, term, ..., reference = NULL, ci = TRUE) {
 
   expressions <- quos(...)
@@ -60,6 +56,8 @@ gg_partial <- function(data, model, term, ..., reference = NULL, ci = TRUE) {
 #' @rdname gg_partial
 #' @inherit gg_partial
 #' @importFrom tidyr complete
+#' @param time_var The name of the variable that was used in \code{model} to
+#' represent follow-up time.
 #' @export
 gg_partial_ll <- function(
   data,
@@ -111,16 +109,32 @@ gg_partial_ll <- function(
 
 }
 
-#' @rdname gg_partial
+#' Plot 1D (smooth) effects
+#'
+#' Flexible, high-level plotting function for (non-linear) effects conditional
+#' on further covariate specifications and potentially relative to
+#' a comparison specification.
+#'
+#' @inheritParams gg_partial
 #' @importFrom purrr map_int
 #' @importFrom rlang quos
+#' @examples
+#' ped <- tumor[1:200, ] %>% as_ped(Surv(days, status) ~ . )
+#' model <- mgcv::gam(ped_status~s(tend) + s(age, by = complications), data=ped,
+#'   family = poisson(), offset=offset)
+#' make_newdata(ped, age = seq_range(age, 20), complications = levels(complications))
+#' gg_slice(ped, model, "age", age=seq_range(age, 20), complications=levels(complications))
+#' gg_slice(ped, model, "age", age=seq_range(age, 20), complications=levels(complications),
+#'  ci = FALSE)
+#' gg_slice(ped, model, "age", age=seq_range(age, 20), complications=levels(complications),
+#'   reference=list(age = 50))
 #' @export
-gg_slice <- function(data, model, term, ..., reference=NULL) {
+gg_slice <- function(data, model, term, ..., reference = NULL, ci = TRUE) {
 
   expressions <- quos(...)
   vars        <- names(expressions)
   ndf         <- make_newdata(data, ...) %>%
-    add_term2(model, term, reference = reference)
+    add_term2(model, term, reference = reference, se.fit = ci)
 
   n_unique <- map_int(vars, ~length(unique(ndf[[.x]])))
   vars     <- vars[rev(order(n_unique))]
@@ -128,12 +142,20 @@ gg_slice <- function(data, model, term, ..., reference=NULL) {
   ndf      <- ndf %>% mutate_at(vars[-1], ~as.factor(.x))
   n_vars   <- length(vars)
 
-  gg_out <- ggplot(ndf, aes_string(x = vars[1], y = "fit")) +
-    geom_ribbon(aes_string(ymin = "ci_lower", ymax = "ci_upper"), alpha = 0.3) +
-    geom_line()
+  gg_out <- ggplot(ndf, aes_string(x = vars[1], y = "fit"))
+  if (ci) {
+    gg_out <- gg_out +
+      geom_ribbon(aes_string(ymin = "ci_lower", ymax = "ci_upper"), alpha = 0.3)
+  }
+  gg_out <- gg_out + geom_line()
   if (n_vars > 1) {
-    gg_out <- gg_out + aes_string(group = vars[2], fill = vars[2]) +
+    if(ci) {
+      gg_out <- gg_out + aes_string(group = vars[2], fill = vars[2]) +
       geom_line(aes_string(col = vars[2]))
+    } else {
+      gg_out <- gg_out + aes_string(group = vars[2]) +
+        geom_line(aes_string(col = vars[2]))
+    }
     if (n_vars > 2) {
       form   <- as.formula(paste0("~", vars[-1:-2], collapse = "+"))
       gg_out <- gg_out + facet_wrap(form, labeller = label_both)
@@ -145,41 +167,45 @@ gg_slice <- function(data, model, term, ..., reference=NULL) {
 }
 
 
-#' @rdname gg_partial
-#' @inherit gg_partial
-#'
+#' @rdname get_cumu_eff
+#' @inherit get_cumu_eff
+#' @inheritParams get_cumu_eff
+#' @inheritParams gg_partial
 #' @export
-gg_cumu_eff <- function(data, model, term, z1, z2=NULL, se_mult = 2) {
+gg_cumu_eff <- function(data, model, term, z1, z2=NULL, se_mult = 2, ci = TRUE) {
 
   cumu_eff_df <- get_cumu_eff(data, model, term, z1, z2, se_mult)
 
-  ggplot(cumu_eff_df, aes_string(x = "tend", y = "cumu_eff")) +
-    geom_ribbon(aes_string(ymin = "cumu_eff_lower", ymax = "cumu_eff_upper"),
-      alpha = 0.3) +
-    geom_line() +
-    xlab("time") + ylab("cumulative effect")
+  gg_out <- ggplot(cumu_eff_df, aes_string(x = "tend", y = "cumu_eff"))
+  if (ci) {
+    gg_out <- gg_out +
+      geom_ribbon(aes_string(ymin = "cumu_eff_lower", ymax = "cumu_eff_upper"),
+        alpha = 0.3)
+  }
+
+  gg_out + geom_line() + xlab("time") + ylab("cumulative effect")
 
 }
 
 
 
 
-#' @inherit gg_partial
+#' @inherit gg_partial_ll
+#' @rdname gg_partial
 #' @export
-#' @keywords internal
 get_partial_ll <- function(
-  x,
+  data,
   model,
   term, ...,
   reference = NULL,
   ci        = FALSE,
   time_var  = "tend") {
 
-  ind_term <- which(map_lgl(attr(x, "func_mat_names"), ~any(grepl(term, .x))))
-  tz_var   <- attr(x, "tz_vars")[[ind_term]]
-  tz_val   <- attr(x, "tz")[[ind_term]]
+  ind_term <- which(map_lgl(attr(data, "func_mat_names"), ~any(grepl(term, .x))))
+  tz_var   <- attr(data, "tz_vars")[[ind_term]]
+  tz_val   <- attr(data, "tz")[[ind_term]]
 
-  ll_df <- get_ll(x, ind_term, ..., time_var = time_var)  %>%
+  ll_df <- get_ll(data, ind_term, ..., time_var = time_var)  %>%
     add_term2(object = model, term = term, reference = reference, se.fit = ci)
 
 }